+++ package.json
@@ -23,6 +23,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
+    "node-cron": "^4.1.1",
     "nodemon": "^3.1.10",
     "pg": "^8.16.2"
   }
