+++ routes/authRoutes.js
@@ -1,69 +1,15 @@
 import express from "express";
-import bcrypt, { hash } from "bcrypt";
-import connections from "../db/dbConnection.js";
+import authRoutesHandler from "../routesHandler/authRoutesHandler.js";
 import {config} from "dotenv";
+import getRemainingLeave from "../hooks/getRemainingLeave.js";
 config();
 const authRouter = express.Router();
 
-
-authRouter.post("/login",async(req,res)=>{
-    const data = req.body;
-    if(!data){
-        return res.json({message:"Error!Cannot get the data"});
-    }
-    if(!data.phone.trim() ||!data.password.trim()){
-        return res.json({message:"Fill all the fields"})
-    }
-    const result = await connections.query(`Select password from users where id='${data.phone}';`);
-    const password = result.rows[0]?.password;
-    if(!password) return res.json({message:"Phone or Password Invalid"});
-    if(password===process.env.DEFAULTPASSWORD && data.password===password){
-        let result = await connections.query(`
-                UPDATE users 
-                SET allowChangePassword = 'true' 
-                WHERE id = '${data.phone}'`);;
-        return res.json({message:"Default password detected,Kindly change the password.You can change the password by logging in with the password you wish to change"});
-        
-    }
-    else if(password===process.env.DEFAULTPASSWORD && data.password !== password){
-        let result = await connections.query(`select allowChangePassword from users where id=${data.phone}`);
-        let canChange = result.rows[0]?.allowchangepassword;
-        console.log(canChange);
-        if(canChange==='true'){
-            const salt =bcrypt.genSaltSync(Number(process.env.bcryptKey));
-            const hashed = bcrypt.hashSync(data.password,salt);
-
-            result = await connections.query(`update users 
-                SET password='${hashed}',
-                allowchangepassword='false'
-                 where id='${data.phone}'`);
-            return res.json({message:"Changed Password"});
-        }
-        return res.json({message:"Wrong Password"});
-    }
-    else{
-    const isCorrect = bcrypt.compareSync(data.password,password);
-    if(isCorrect){
-       return res.json({message:"welcome user"});
-    }
-    res.json({message:"Phone or Password Invalid"});}
-
-})
-
-authRouter.post("/startwork",async(req,res)=>{
-    const data = req.body;
-    if(!data) return res.json({message:"Data not Fetched"});
-    if(!data.id)return res.json({message:"id cannot be fetched"});
-    console.log(data.id);
-    try{
-        const result = await connections.query(`insert into attendance(user_id) Values(${data.id});`);
-    }
-    catch(e){
-        console.log(e);
-        return res.json({message:"Error starting work"});
-    }
-    res.json({message:"Successfully started work"});
-
-})
+//to login
+authRouter.post("/login",authRoutesHandler.loginHandler);
+//to start work
+authRouter.post("/startwork",authRoutesHandler.startWorkHandler);
+//to see the remaining weekly leaves available
+authRouter.get("/getRemainingLeave/:id",authRoutesHandler.getRemaining);
 
 export default authRouter;
